Moved some files and rewrote some  parts of the code
