+++ src/app/(details)/about/card.tsx
