+++ src/app/(details)/about/page.tsx
@@ -1,8 +1,6 @@
 import { Metadata } from 'next'
 import AboutPage from '@/components/staticPages/about-page'
 
-import ProjectData from '@/data/projects.json'
-
 export const metadata: Metadata = {
   title: 'About',
   description: 'The about page of my personal website',
@@ -15,12 +13,6 @@ export const metadata: Metadata = {
   },
 }
 
-async function getHighlighted() {
-  return ProjectData
-}
-
 export default async function About() {
-  const highlightedProjects = await getHighlighted()
-
-  return <AboutPage highlightedProjects={highlightedProjects} />
+  return <AboutPage />
 }
