+++ src/app/(details)/layout.tsx
@@ -1,0 +1,3 @@
+export default function DetailsLayout({ children }: any) {
+  return <section className="md:px-0 py-12 md:py-32">{children}</section>
+}
