+++ src/app/(details)/projects/[slug]/loading.tsx
