+++ src/app/(details)/projects/[slug]/not-found.tsx
