+++ src/app/(details)/projects/[slug]/page.tsx
