+++ src/components/ui/elongated-image.tsx
@@ -1,4 +1,5 @@
 'use client'
+
 import Image from 'next/image'
 import { useEffect, useState } from 'react'
 import { ElongatedImageProps as Props } from '@/types/props'
